Move content and JS files to new `src` folder
