+++ .eleventyignore
@@ -1,1 +1,1 @@
-./content/episodes/_template.md
+./src/content/episodes/_template.md
