+++ src/js/components/MaxWidth.js
